// controls/richtexteditor/blazor-script/editor-scripts/common/interface.ts
export interface IMarkdownFormatterCallBack {
    selectedText?: string
    editorMode?: EditorMode
    action?: string
    event?: KeyboardEvent | MouseEvent
    requestType?: string
}

/**
 * @deprecated
 */
export interface IHtmlFormatterCallBack {
    selectedNode?: Element
    requestType?: string
    range?: Range
    editorMode?: EditorMode
    action?: string
    elements?: Element | Element[]
    imgElem?: Element | Element[]
    event?: KeyboardEvent | MouseEvent | ClipboardEvent
    isKeyboardEvent?: boolean
}

/**
 * @deprecated
 */
export interface IMarkdownToolbarStatus {
    OrderedList: boolean
    UnorderedList: boolean
    Formats: string
}
/**
 * @deprecated
 */
export interface IUndoCallBack {
    callBack?: Function
    event?: Object
}
// ...
export interface IEditorModel {
    currentDocument?: Document
    execCommand?: Function
    observer?: Observer
    markdownSelection?: MarkdownSelection
    undoRedoManager?: UndoRedoManager | UndoRedoCommands
    nodeSelection?: NodeSelection
    mdSelectionFormats?: MDSelectionFormats
    domNode?: DOMNode
    nodeCutter?: NodeCutter
    formatPainterEditor?: IFormatPainterEditor
    domTree?: DOMMethods
    linkObj?: LinkCommand | MDLink
    videoObj?: VideoCommand
    audioObj?: AudioCommand
    imgObj?: ImageCommand
    formatPainterObj?: FormatPainterActions
    tableObj?: TableCommand | MDTable
    pasteObj?: PasteCleanupAction
    emojiPickerObj?: EmojiPickerAction
    editableElement?: Element
    userAgentData?: CustomUserAgentData
    destroy?(): void
    beforeSlashMenuApplyFormat?(): void
    codeBlockObj?: CodeBlockPlugin
}
/**
 * Interface for checklist item model
 */
export interface IChecklistItemModel {
    /** Defines the type of the checklist item */
    text: string;
    iconCss: string;
}
/**
 * @hidden
 * @deprecated
 */
export interface IDropDownItemModel extends DropDownItemModel {
    cssClass?: string
    command?: string
    subCommand?: string
    value?: string
    text?: string
}

/**
 * @hidden
 * @deprecated
 */
export interface IToolbarItemModel extends ItemModel {
    command?: string
    subCommand?: string
}
// ...
export interface IVideoCommandsArgs {
    /** Specifies the source URL of the video. */
    url?: string
    /** Represents the instance of the current selection within the editor. */
    selection?: NodeSelection
    /** Defines the minimum, maximum, and current width of the video. */
    width?: { minWidth?: string | number; maxWidth?: string | number; width?: string | number }
    /** Defines the minimum, maximum, and current height of the video. */
    height?: { minHeight?: string | number; maxHeight?: string | number; height?: string | number }
    /** Specifies the file name of the video, which can be a string or a DocumentFragment. */
    fileName?: string | DocumentFragment
    /** Indicates whether the video link is an embedded URL. */
    isEmbedUrl?: boolean
    /** Specifies the CSS class to be applied to the video element. */
    cssClass?: string
    /** Represents the selected parent node of the video element to be edited. */
    selectParent?: Node[]
    /** Specifies the title attribute for the video element. */
    title?: string
}
// ...
export interface ActionBeginEventArgs {
    /** Specifies the type of the current action. */
    requestType?: string
    /** Indicates whether to cancel the current action. */
    cancel?: boolean
    /**
     * Specifies the current toolbar or dropdown item involved in the action.
     *
     * @deprecated
     */
    item?: IToolbarItemModel | IDropDownItemModel
    /** Specifies the event that initiated the action, such as mouse, keyboard, or drag events. */
    originalEvent?: MouseEvent | KeyboardEvent | DragEvent
    /** Specifies the name of the event. */
    name?: string
    /** Specifies whether the selection type is a dropdown. */
    selectType?: string
    /**
     * Provides details about URL actions.
     *
     * @deprecated
     */
    itemCollection?: IItemCollectionArgs
    /**
     * Defines the emoji picker details.
     *
     * @deprecated
     */
    emojiPickerArgs?: IEmojiPickerArgs
    /**
     * Defines the content to be exported.
     *
     * @deprecated
     */
    exportValue?: string
}

export interface IEmojiPickerArgs {
    emojiSettings: EmojiSettingsModel

}

/**
 * Provides detailed information about a Print event in the Rich Text Editor (RTE).
 */
export interface PrintEventArgs extends ActionBeginEventArgs {
    /** Defines the Rich Text Editor (RTE) element associated with the Print event. */
    element?: Element
}

/**
 * @deprecated
 */
export interface IItemCollectionArgs {
    /** Defines the instance of the current selection */
    selection?: NodeSelection
    /** Defines the HTML elements of currently selected content */
    selectNode?: Node[]
    /** Defines the parent HTML elements of current selection */
    selectParent?: Node[]
    /** Defines the URL action details for link element */
    url?: string
    /** Defines the Display Text action details for link element */
    text?: string
    /** Defines the title of the link action details */
    title?: string
    /** Defines the target as string for link element */
    target?: string
    /** Defines the element to be inserted */
    insertElement?: Element
}
// ...
export interface ActionCompleteEventArgs {
    /** Specifies the type of the current action. */
    requestType?: string
    /** Specifies the name of the event. */
    name?: string
    /** Specifies the current mode of the editor. */
    editorMode?: string
    /**
     * Defines the selected elements.
     *
     * @deprecated
     */
    elements?: Node[];
    /** Specifies the event associated with the action, such as a mouse or keyboard event. */
    event?: MouseEvent | KeyboardEvent;
    /**
     * Defines the selected range.
     *
     * @deprecated
     */
    range?: Range
}
// ...
export interface IShowPopupArgs {
    args?: MouseEvent | TouchEvent | KeyboardEvent
    type?: string
    isNotify: boolean
    elements?: Element | Element[]
}